requirements.txt
paho-mqtt>=2.1.0
requests>=2.32.5
meshcoredecoder>=0.1.0